resolves PR related issues; tests use Parametrized runner;
