+++ extended/src/main/java/apoc/ml/OpenAI.java
@@ -35,7 +35,7 @@ public class OpenAI {
     public static final String APIKEY_CONF_KEY = "apiKey";
     public static final String JSON_PATH_CONF_KEY = "jsonPath";
     public static final String PATH_CONF_KEY = "path";
-    public static final String GPT_4O_MODEL = "gpt-4.1";
+    public static final String GPT_4O_MODEL = "gpt-4o";
     public static final String FAIL_ON_ERROR_CONF = "failOnError";
     public static final String ENABLE_BACK_OFF_RETRIES_CONF_KEY = "enableBackOffRetries";
     public static final String ENABLE_EXPONENTIAL_BACK_OFF_CONF_KEY = "exponentialBackoff";
@@ -73,7 +73,7 @@ public class OpenAI {
                 apocConfig.getString(APOC_ML_OPENAI_TYPE, OpenAIRequestHandler.Type.OPENAI.name())
         );
         OpenAIRequestHandler.Type type = OpenAIRequestHandler.Type.valueOf(apiTypeString.toUpperCase(Locale.ENGLISH));
-        
+
         var configForPayload = new HashMap<>(configuration);
         // we remove these keys from configPayload, since the json payload is calculated starting from the configPayload map
         Stream.of(ENDPOINT_CONF_KEY, API_TYPE_CONF_KEY, API_VERSION_CONF_KEY, APIKEY_CONF_KEY).forEach(configForPayload::remove);
@@ -90,7 +90,7 @@ public class OpenAI {
         apiType.addApiKey(headers, apiKey);
 
         String payload = JsonUtil.OBJECT_MAPPER.writeValueAsString(configForPayload);
-        
+
         // new URL(endpoint), path) can produce a wrong path, since endpoint can have for example embedding,
         // eg: https://my-resource.openai.azure.com/openai/deployments/apoc-embeddings-model
         // therefore is better to join the not-empty path pieces
@@ -133,7 +133,7 @@ public class OpenAI {
                 } else {
                     configuration.putIfAbsent(PATH_CONF_KEY, "messages");
                     configForPayload.putIfAbsent(MAX_TOKENS, 1000);
-                    configForPayload.putIfAbsent(MODEL_CONF_KEY, "claude-3-7-sonnet-20250219");
+                    configForPayload.putIfAbsent(MODEL_CONF_KEY, "claude-3-5-sonnet-20240620");
                 }
 
                 configForPayload.remove(ANTHROPIC_VERSION);
@@ -170,7 +170,7 @@ public class OpenAI {
                 (map, text) -> {
                     Long index = (Long) map.get("index");
                     return new EmbeddingResult(index, text, (List<Double>) map.get("embedding"));
-                }, 
+                },
                 m -> new EmbeddingResult(-1, m, List.of())
         );
     }
@@ -181,7 +181,7 @@ public class OpenAI {
         if (texts == null) {
             throw new RuntimeException(ERROR_NULL_INPUT);
         }
-        
+
         Map<Boolean, List<String>> collect = texts.stream()
                 .collect(Collectors.groupingBy(Objects::nonNull));
