+++ extended/src/test/java/apoc/ml/OpenAiAzureIT.java
@@ -6,9 +6,13 @@ import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Ignore;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 import org.neo4j.test.rule.DbmsRule;
 import org.neo4j.test.rule.ImpermanentDbmsRule;
 
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.Map;
 import java.util.stream.Stream;
 
@@ -23,7 +27,8 @@ import static apoc.ml.OpenAITestResultUtils.assertCompletion;
 import static apoc.util.TestUtil.testCall;
 import static org.junit.Assume.assumeNotNull;
 
-public abstract class OpenAiAzureBaseIT {
+@RunWith(Parameterized.class)
+public class OpenAiAzureIT {
     // In Azure, the endpoints can be different 
     private static String OPENAI_EMBEDDING_URL;
     private static String OPENAI_CHAT_URL;
@@ -36,7 +41,20 @@ public abstract class OpenAiAzureBaseIT {
     @ClassRule
     public static DbmsRule db = new ImpermanentDbmsRule();
 
-    abstract String getDefModel();
+    @Parameterized.Parameters(name = "chatModel: {0}")
+    public static Collection<String[]> data() {
+        return Arrays.asList(new String[][] {
+                // tests with model evaluated
+                {"gpt-35-turbo"},
+                {"gpt-4.1"},
+                {"gpt-4o"},
+                // tests with default model
+                {null}
+        });
+    }
+
+    @Parameterized.Parameter(0)
+    public String chatModel;
 
     @BeforeClass
     public static void setUp() throws Exception {
@@ -86,13 +104,13 @@ public abstract class OpenAiAzureBaseIT {
     public void completion() {
         testCall(db, COMPLETION_QUERY,
                 getParams(OPENAI_CHAT_URL),
-                (row) -> assertCompletion(row, getDefModel()));
+                (row) -> assertCompletion(row, chatModel));
     }
 
     @Test
     public void chatCompletion() {
         testCall(db, CHAT_COMPLETION_QUERY, getParams(OPENAI_COMPLETION_URL),
-                (row) -> assertChatCompletion(row, getDefModel()));
+                (row) -> assertChatCompletion(row, chatModel));
     }
 
     private Map<String, Object> getParams(String url) {
@@ -101,7 +119,7 @@ public abstract class OpenAiAzureBaseIT {
                         API_TYPE_CONF_KEY, OpenAIRequestHandler.Type.AZURE.name(),
                         API_VERSION_CONF_KEY, OPENAI_AZURE_API_VERSION,
                         // on Azure is available only "gpt-35-turbo"
-                        "model", getDefModel()
+                        "model",chatModel
                 )
         );
     }
